No more warnings

+++ main.rs
@@ -1,5 +1,5 @@
 #![no_std]
-#![allow(ctypes)]
+#![allow(improper_ctypes)]
 
 #![feature(lang_items)]
 #[lang="sized"]
@@ -10,6 +10,7 @@ trait Copy {}
 
 impl Copy for Color {}
 
+#[allow(dead_code)]
 enum Color {
     Black      = 0,
     Blue       = 1,
@@ -35,12 +36,12 @@ enum Option<T> {
 }
 
 struct IntRange {
-    cur: int,
-    max: int
+    cur: isize,
+    max: isize
 }
 
 impl IntRange {
-    fn next(&mut self) -> Option<int> {
+    fn next(&mut self) -> Option<isize> {
         if self.cur < self.max {
             self.cur += 1;
             Option::Some(self.cur - 1)
@@ -50,7 +51,7 @@ impl IntRange {
     }
 }
 
-fn range(lo: int, hi: int) -> IntRange {
+fn range(lo: isize, hi: isize) -> IntRange {
     IntRange { cur: lo, max: hi }
 }
 
@@ -69,7 +70,7 @@ fn clear_screen(background: Color) {
 }
 
 #[no_mangle]
-#[no_split_stack]
+#[no_stack_check]
 pub fn main() {
     clear_screen(Color::LightRed);
 }
